merthur

+++ DataStructures/src/edu/algorithms/SelectionSort.java
@@ -2,10 +2,10 @@ package edu.algorithms;
 
 public class SelectionSort {
 	/** The method for sorting the numbers */
-	public static void selectionSort(double[] list) {
+	public static void selectionSort(int[] list) {
 		for (int i = 0; i < list.length - 1; i++) {
 // Find the minimum in the list[i..list.length-1]
-			double currentMin = list[i];
+			int currentMin = list[i];
 			int currentMinIndex = i;
 			for (int j = i + 1; j < list.length; j++) {
 				if (currentMin > list[j]) {
